Scope DINO loss to HiDream-O1 trainer
